Rename domain to dto
