+++ security-server/security-server-dto/pom.xml
@@ -9,7 +9,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>security-server-domain</artifactId>
+    <artifactId>security-server-dto</artifactId>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
