+++ security-server/security-server-rest/src/main/java/by/kovzov/uis/security/rest/security/service/TokenService.java
@@ -2,7 +2,7 @@ package by.kovzov.uis.security.rest.security.service;
 
 import org.springframework.security.core.Authentication;
 
-import by.kovzov.uis.security.domain.dto.JwtAuthenticationDto;
+import by.kovzov.uis.security.dto.JwtAuthenticationDto;
 
 public interface TokenService {
