+++ security-server/security-server-rest/src/main/java/by/kovzov/uis/security/rest/security/service/TokenServiceImpl.java
@@ -16,7 +16,7 @@ import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 
-import by.kovzov.uis.security.domain.dto.JwtAuthenticationDto;
+import by.kovzov.uis.security.dto.JwtAuthenticationDto;
 import by.kovzov.uis.security.rest.security.model.UserSecurity;
 import lombok.RequiredArgsConstructor;
